pyproject.toml
[tool.poetry]
name = "pivobot"
version = "0.1.0"
description = ""
authors = ["Mark Nicholson <i@justmark0.me>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.21"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
